+++ FrontEnd/src/App.js
@@ -78,7 +78,7 @@ function App () {
     const [workflow, setWorkflow] = useState()
  
     // Custom hooks
-    const { message, files, error: messageError, isProcessing, handleSubmit } = useSubmitMessage(concatenatedQA, selectedFiles, selectedMessages, tags, workflow, setWorkflow);
+    const { message, files, error: messageError, isProcessing, handleSubmit, disconnectFromRequest } = useSubmitMessage(concatenatedQA, selectedFiles, selectedMessages, tags, workflow, setWorkflow);
     const { augmentedPrompt, setAugmentedPrompt, isAugmenting, error: augmentedError, generateAugmentedPrompt } = useAugmentedPrompt();
     const { questionsForPrompt, setQuestionsForPrompt, isQuestioning, error: questionsError, generateQuestionsForPrompt } = useSuggestedQuestions();
     const { selectedWorkflow, workflowIsLoading, selectMessageError, selectWorkflow } = useSelectedWorkflow();
@@ -216,6 +216,7 @@ function App () {
             {/* ToDo: Should expand out on hover */}
             <UserInputForm 
               handleSubmit={handleFormSubmit}
+              disconnectFromRequest={disconnectFromRequest}
               handleInputChange={handleInputChange}
               userInput={userInput}
               isProcessing={isProcessing}